% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Custom histogram plot}
\usage{
accuracy(data, y, prob, p1, p2 = 1, filter, label1, label2)
}
\arguments{
\item{data}{A data frame}

\item{y}{The variable of interest}

\item{prob}{The predicted probability column}

\item{p1}{The minimun probability to consider label1 true}

\item{p2}{The maximun probability to consider label1 true}

\item{filter}{A label of interest to filter from y}

\item{label1}{First label to compare with y labels}

\item{label2}{Second label to compare with y labels}
}
\description{
This function will create a custom histogram
}
\examples{
accuracy(data = data, y = "y", prob = "x", p1 = 0.5, filter = "label1", label1 = "label1", label2 = "label2")

}
