% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_prop_plot.R
\name{bar_prop_plot}
\alias{bar_prop_plot}
\title{Custom proportion bar plot}
\usage{
bar_prop_plot(data, x, var, xlab = "", ylab = "", title = "")
}
\arguments{
\item{data}{A data frame}

\item{x}{X axis variable}

\item{var}{Filter variable}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{title}{boxplot title}
}
\description{
This function will create a custom proportion bar plot
}
\examples{
bar_prop_plot(data, Gender, IsWorking)

}
