% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{bxplt}
\alias{bxplt}
\title{Custom boxplot}
\usage{
bxplt(data, x, y, xlab = "", ylab = "", title = "")
}
\arguments{
\item{data}{A data frame}

\item{x}{A variable}

\item{y}{Another variable}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{title}{boxplot title}
}
\description{
This function will create a custom boxplot
}
\examples{
bxplt(data, "Gender", "Age")

}
