% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_table.R
\name{coef_table}
\alias{coef_table}
\title{Custom markdown table}
\usage{
coef_table(model, title = "")
}
\arguments{
\item{title}{Title of the table}

\item{dmodel}{A logit model}
}
\description{
This function will create a custom model coefficients table
}
\examples{
coef_table(model)

}
