% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_proptable.R
\name{cont_prop.table}
\alias{cont_prop.table}
\title{Creates a contingency table}
\usage{
cont_prop.table(col1, col2)
}
\arguments{
\item{col1}{A data frame and it's first column}

\item{col2}{A data frame and it's second column}
}
\description{
This function will return a contingency table of column proportion from 2 discrete variables
}
\examples{
cont_prop.table(data$variable1, data$variable2)

}
