% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Custom histogram plot}
\usage{
histogram(data, var, x, filter, xlab = "", ylab = "", title = "")
}
\arguments{
\item{data}{A data frame}

\item{var}{A variable to the data be filtered from}

\item{x}{The variable to build the histogram}

\item{filter}{The label to filter the var parameter}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{title}{histogram title}
}
\description{
This function will create a custom histogram
}
\examples{
histogram(data, "Gender", "Age", "Female")

}
