% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_model.R
\name{logit_model}
\alias{logit_model}
\title{Logit model}
\usage{
logit_model(data, formula)
}
\arguments{
\item{data}{A data frame to train the model}

\item{formula}{A data frame and it's second column}
}
\description{
This function will return a logit model
}
\examples{
logit_model(data, y ~ x1 + ... + xn)

}
