% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_predict.R
\name{logit_predict}
\alias{logit_predict}
\title{Model predictions}
\usage{
logit_predict(data, model)
}
\arguments{
\item{data}{A data frame}

\item{model}{A model}
}
\description{
This function will return the data frame with an extra column, which is the predictions
}
\examples{
logit_predict(data, y ~ x1 + ... + xn)

}
