% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_table.R
\name{rmd_table}
\alias{rmd_table}
\title{Custom markdown table}
\usage{
rmd_table(data, title = "")
}
\arguments{
\item{title}{Title of the table}

\item{datavar}{A data frame}
}
\description{
This function will create a custom table
}
\examples{
summary_table(data$Gender)

}
