% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumNA.R
\name{sum_na}
\alias{sum_na}
\title{Sum missing data}
\usage{
sum_na(variable)
}
\arguments{
\item{variable}{A data frame and it's column}
}
\description{
This function will return the number of missing data on a certain column of a data frame
}
\examples{
sum_na(data$Gender)

}
