% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_filtered.R
\name{summary_filtered}
\alias{summary_filtered}
\title{Custom summary data frame}
\usage{
summary_filtered(data, var, x, row.names)
}
\arguments{
\item{data}{A data frame}

\item{var}{A variable to the data be filtered from}

\item{x}{The variable of interest}

\item{row.names}{A vector with row names ordered}
}
\description{
This function will create a summary data frame by categories
}
\examples{
summary_filtered(data, Gender, "Age", c("Male","Female"))

}
