% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Custom summary data frame}
\usage{
summary_table(datavar)
}
\arguments{
\item{datavar}{A data frame and it's column of interest}
}
\description{
This function will create a summary table
}
\examples{
summary_table(data$Gender)

}
