% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treat}
\alias{treat}
\title{Decode treatment function}
\usage{
treat(data, col, newvar, code1, code2, decode1, decode2)
}
\arguments{
\item{data}{A data frame}

\item{col}{A column you wish to decode}

\item{newvar}{The name of the new column (it can be the same name of the original variable)}

\item{code1}{The first label in use}

\item{code2}{The second label in use}

\item{decode1}{The label you want to replace the first coded label}

\item{decode2}{The label to replace the remaining coded label}
}
\description{
This function can decode a 2 labels variable. The variable must be character
}
\examples{
treat(data, Gender, "Gender", "F", "M", "Female", "Male")

}
